Support DLA_NbrLockBits on 256 bits and Rsvd_Area_Size on 256 bytes
Bug: 169640640
Test: Tag R/W

+++ SN100x/src/nfc/include/rw_int.h
@@ -386,14 +386,14 @@ typedef uint8_t tRW_T2T_LOCK_STATUS;
 
 typedef struct {
   uint16_t offset;              /* Offset of the lock byte in the Tag */
-  uint8_t num_bits;             /* Number of lock bits in the lock byte */
+  uint16_t num_bits;            /* Number of lock bits in the lock byte */
   uint8_t bytes_locked_per_bit; /* No. of tag bytes gets locked by a bit in this
                                    byte       */
 } tRW_T2T_LOCK_INFO;
 
 typedef struct {
   uint16_t offset;   /* Reserved bytes offset taken from Memory control TLV */
-  uint8_t num_bytes; /* Number of reserved bytes as per the TLV */
+  uint16_t num_bytes; /* Number of reserved bytes as per the TLV */
 } tRW_T2T_RES_INFO;
 
 typedef struct {
